seg_sz is a WORD not a DWORD
I had the wrong size for the segment size. This commit corrects it and
passes seg_sz as a parameter to writeSegment so that the output should
be directly comparable to the output of the master branch.

+++ qic122.c
@@ -24,9 +24,9 @@ unsigned int decompressExtent(BYTE *cbuf, BYTE *dbuf) {
     unsigned int frame = 0;
 
     /*
-     * Treat the first two bytes of cbuf as a DWORD segment size
+     * Treat the first two bytes of cbuf as a WORD segment size
      */
-    while((comp_sz = (DWORD)cbuf[comp_rd] | (DWORD)(cbuf[comp_rd + 1] << 8)) != 0) {
+    while((comp_sz = (WORD)cbuf[comp_rd] | (WORD)(cbuf[comp_rd + 1] << 8)) != 0) {
 
         comp_rd += comp_sz + 2;
